requirements.txt
django==4.2.0
langchain-core==0.1.25
langchain>=0.1.5
langchain-openai>=0.0.5
langchain-community>=0.0.17,<0.1
langgraph==0.0.27
openai==1.12.0
#ncloud-sdk==1.1.9
djangorestframework==3.14.0
channels==4.0.0
whitenoise==6.6.0
cryptography==41.0.7
#faiss-cpu==1.7.4
django-allauth==0.58.2
python-dotenv==1.0.0
daphne==4.0.0
django-cors-headers==4.3.0
gunicorn==21.2.0
pydantic==2.5.3